+++ app/Http/Controllers/OccupationController.php
@@ -7,18 +7,25 @@ use App\Models\Occupation;
 use App\Models\Department;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\OccupationRequest;
 
 class OccupationController extends Controller
 {
     public function index()
     {
+        $occupation = new Occupation();
+        Gate::authorize('list', $occupation);
+
         $occupations = Occupation::query()->latest('id')->get();
         return view('occupation.index',['occupations' => $occupations]);
     }
 
     public function create()
     {
+        $occupation = new Occupation();
+        Gate::authorize('crud', $occupation);
+
         $departments = Department::query()->latest('id')->get();
         return view('occupation.form',[
             'occupation' => new Occupation(),
@@ -33,10 +40,14 @@ class OccupationController extends Controller
 
     public function store(OccupationRequest $request)
     {
+        $occupation = new Occupation();
+        Gate::authorize('crud', $occupation);
+
         $occupation = Occupation::create([
             'name' => $request->validated('name'),
             'department_id' => $request->validated('department'),
         ]);
+        
         Log::create(['user_id' => auth()->user()->id, 'email' => auth()->user()->email, 'log' => 'created occupation_id - '.$occupation->id]);
         return to_route('occupations.index');
     }
@@ -48,6 +59,8 @@ class OccupationController extends Controller
 
     public function edit(Occupation $occupation)
     {
+        Gate::authorize('crud', $occupation);
+
         $departments = Department::query()->latest('id')->get();
         return view('occupation.form',[
             'occupation' => $occupation,
@@ -62,17 +75,23 @@ class OccupationController extends Controller
 
     public function update(OccupationRequest $request, Occupation $occupation)
     {
+        Gate::authorize('crud', $occupation);
+
         $occupation->update([
             'name' => $request->name,
             'department_id' => $request->department,
         ]);
+        
         Log::create(['user_id' => auth()->user()->id, 'email' => auth()->user()->email, 'log' => 'updated occupation_id - '.$occupation->id]);
         return to_route('occupations.index');
     }
 
     public function destroy(Occupation $occupation)
     {
+        Gate::authorize('crud', $occupation);
+
         Occupation::query()->where('id', $occupation->id)->delete();
+        
         Log::create(['user_id' => auth()->user()->id, 'email' => auth()->user()->email, 'log' => 'deleted occupation_id - '.$occupation->id]);
         return to_route('occupations.index');
     }
